pyproject.toml
[project]
name = "syft-llm-router"
version = "0.1.0"
description = "Syft LLM Router Protocol"
requires-python = ">=3.9"
readme = "README.md"
dependencies = [
    "syft-core>=0.2.7",
    "fastsyftbox>=0.1.19",
    "ipython>=8.18.1",
    "jinja2>=3.1.6",
    "pydantic>=2.11.2",
    "requests>=2.32.3",
    "sqlmodel>=0.0.24",
    "typer>=0.15.2",
    "python-dotenv>=1.1.1",
    "syft-accounting-sdk @ git+https://github.com/OpenMined/accounting-sdk.git"
]



[tool.uv]
dev-dependencies = [
    "openai>=1.70.0",
]
